feat: move code structure to a more modular one
