+++ src/hermes/hermes.types.ts
@@ -5,30 +5,10 @@ import {
   RecognizedString
 } from 'uWebSockets.js'
 
-export enum RequestMethods {
-  GET = 'GET',
-  POST = 'POST',
-  PUT = 'PUT',
-  DELETE = 'DELETE'
-}
-
+/*-- Main Definitions */
 export interface MainDependencies {
   logger: any | Console
 }
-
-export interface RouterResponse {
-  status: number,
-  message?: string,
-  responseType?: 'json' | 'text'
-}
-export interface RouterFunction {
-  (deps: MainDependencies, req: RequestData): Promise<RouterResponse> | RouterResponse
-}
-
-export interface GenAppErrorFunction {
-  (status: RecognizedString, message: string, error?: Error): HermesError
-}
-
 export interface BootstrapFunction {
   (deps: MainDependencies, router: RouterFunction, port?: number, app?: TemplatedApp): void
 }
@@ -38,17 +18,10 @@ export interface RequestHandlerFunction {
 export interface BoundRequestHandler {
   (res: HttpResponse, req: HttpRequest): Promise<HttpResponse>
 }
-export interface ErrorHandler {
-  (res: HttpResponse, logger: any | Console, error: HermesError): HttpResponse
-}
-export interface BoundErrorHandler {
-  (error: HermesError): void
-}
 
 export interface RequestBody {
   [key: string]: string | number
 }
-
 export interface RequestData {
   method: RequestMethods
   route: string
@@ -57,6 +30,32 @@ export interface RequestData {
   headers: Map<string, string>
 }
 
+/*-- Router Definitions */
+export enum RequestMethods {
+  GET = 'GET',
+  POST = 'POST',
+  PUT = 'PUT',
+  DELETE = 'DELETE'
+}
+export interface RouterResponse {
+  status: number,
+  message?: string,
+  responseType?: 'json' | 'text'
+}
+export interface RouterFunction {
+  (deps: MainDependencies, req: RequestData): Promise<RouterResponse> | RouterResponse
+}
+
+/*-- Error Definitions --*/
+export interface GenAppErrorFunction {
+  (status: RecognizedString, message: string, error?: Error): HermesError
+}
+export interface ErrorHandler {
+  (res: HttpResponse, logger: any | Console, error: HermesError): HttpResponse
+}
+export interface BoundErrorHandler {
+  (error: HermesError): void
+}
 export interface HermesError {
   status?: RecognizedString
   message?: string
